Plejady.Accounts (Plejády v1.0.0)

Provides logic for managing accounts. This includes users, guests, admins, etc.

Functions

create_guest(params)

Creates a new guest.

This is a database action.

delete_suggested_admin(id)

Deletes a suggested admin by its id.

This is a database action.

first_user?()

Checks if there are any other users registered. If not, this function returns true.

get_admins()

Fetches all admins.

get_free_guest_places()

Fetches a number of free guest places.

get_user_by_email(email)

Fetches a user by email.

guests()

Fetches all guests.

This is a database action.

strip_off_admin(id)

Strips off admin rights from the user with the given id.

This is a database action.

transfer_admin(me, to_email)

Transfers the admin lead role to another user (specified by their email).

This is a database action.