Plejady.Accounts.SuggestedAdmin (Plejády v1.0.0)

Provides logic for suggested admins (users that will automatically become admins).

Functions

changeset(struct, params \\ %{})

Creates a changeset based on the struct and params.

create(params \\ %{})

Creates a new suggested admin and inserts it into the database.

This is a database action.

delete_by_email(email)

Deletes the suggested admin with the given email.

This is a database action.

get()

Returns the list of all suggested admins.

This is a database action.

get_by_email(email)

Returns the suggested admin with the given email.

This is a database action.