Plejady.Config (Plejády v1.0.0)

Provides the model and logic for the application config.

Uses a GenServer to store the config in memory and provide a simple API for accessing and updating it.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get_config()

Fetches the current config.

set_config(new_config)

Updates the current config.

start_link(_)

Starts the Plejady.Config GenServer implementation.