Plejady.Registry (Plejády v1.0.0)

This module is responsible for managing the registry of presentations and their occupancy. It provides a super fast read-write cache for the real-time component of the app with fire-and-forget requests to the database for permanent storage.

The Plejady.Registry struct is considered a source of truth for the app while running. It is prefetched with data from the database to avoid thousands of lookup requests.

Next, it provides utility functions for fetching the occupancy of a presentation.

Types

t()

@type t() :: %Plejady.Registry{
  presentations: [Plejady.Presentation.t()],
  rooms: [Plejady.Room.t()],
  timeblocks: [Plejady.Timeblock.t()]
}

Functions

cache_name()

Returns the statically-defined name of the cache used by the registry.

clear()

Clears the cache.

fetch_occupancy()

Fetches the occupancy of all presentations.

Returns a map where keys are Plejady.Presentation ids and values are the number of signed users.

fetch_signed_up_for(registry, user_id)

Returns a list of Plejady.Presentation ids that a user specified by user_id is signed up for.

get()

Fetches the registry from the cache.

new(presentations, rooms, timeblocks)

Creates a new registry struct.

update(signed_up_for, presentation_id, user_id)

@spec update(
  signed_up_for :: [String.t()],
  presentation_id :: String.t(),
  user_id :: String.t()
) :: {:ok, [String.t()]} | {:error, reason :: :full | atom()}