Plejady.Room (Plejády v1.0.0)

Provides the model and logic for rooms.

Functions

all()

Returns a list of all rooms.

changeset(room, params \\ %{})

Creates a new changeset based on the room struct and params.

create(params \\ %{})

Creates a new room and inserts it into the database.

This is a database action.

delete(id)

Deletes a room by id.

This is a database action.

get!(id)

Fetch a room by id.

update(room, params \\ %{})

Updates a room with the included params in the database.

This is a database action.