PlejadyWeb.CoreComponents (Plejády v1.0.0)

Provides core UI components.

The components in this module use Tailwind CSS, a utility-first CSS framework. See the Tailwind CSS documentation to learn how to customize the generated components in this module.

Icons are provided by heroicons. See icon/1 for usage.

Functions

back(assigns)

Renders a back navigation link.

Examples

<.back navigate={~p"/posts"}>Back to posts</.back>

Attributes

navigate (:any) (required)

Slots

inner_block (required)

button(assigns)

Renders a button.

Examples

<.button>Send!</.button>
<.button phx-click="go" class="ml-2">Send!</.button>

Attributes

type (:string) - Defaults to nil.

class (:string) - Defaults to nil. Global attributes are accepted.

Slots

inner_block (required)

error(assigns)

Generates a generic error message.

Slots

inner_block (required)

flash(assigns)

Renders flash notices.

Examples

<.flash kind={:info} flash={@flash} />
<.flash kind={:info} phx-mounted={show("#flash")}>Welcome Back!</.flash>

Attributes

id (:string) - the optional id of flash container. Defaults to "flash".

flash (:map) - the map of flash messages to display. Defaults to %{}.

title (:string) - Defaults to nil.

kind (:atom) - used for styling and flash lookup.Must be one of :info, or :error.

autoshow (:boolean) - whether to auto show the flash on mount. Defaults to true.

close (:boolean) - whether the flash can be closed. Defaults to true. Global attributes are accepted.

Slots

inner_block - the optional inner block that renders the flash message.

flash_group(assigns)

Shows the flash group with standard titles and content.

Examples

<.flash_group flash={@flash} />

Attributes

flash (:map) (required) - the map of flash messages.

header(assigns)

Renders a header with title.

Attributes

class (:string) - Defaults to nil.

Slots

inner_block (required)

subtitle

actions

hide(js \\ %JS{}, selector)

hide_modal(js \\ %JS{}, id)

icon(assigns)

Renders a Hero Icon.

Hero icons come in three styles – outline, solid, and mini. By default, the outline style is used, but solid an mini may be applied by using the -solid and -mini suffix.

You can customize the size and colors of the icons by setting width, height, and background color classes.

Icons are extracted from your priv/hero_icons directory and bundled within your compiled app.css by the plugin in your assets/tailwind.config.js.

Examples

<.icon name="hero-cake" />
<.icon name="hero-cake-solid" />
<.icon name="hero-cake-mini" />
<.icon name="hero-bolt" class="bg-blue-500 w-10 h-10" />

Attributes

name (:string) (required)

class (:string) - Defaults to nil.

input(assigns)

Renders an input with label and error messages.

A %Phoenix.HTML.Form{} and field name may be passed to the input to build input names and error messages, or all the attributes and errors may be passed explicitly.

Examples

<.input field={@form[:email]} type="email" />
<.input name="my-input" errors={["oh no!"]} />

Attributes

id (:any) - Defaults to nil.

name (:any)

label (:string) - Defaults to nil.

value (:any)

type (:string) - Defaults to "text".

field (Phoenix.HTML.FormField) - a form field struct retrieved from the form, for example: @form[:email].

errors (:list) - Defaults to [].

checked (:boolean) - the checked flag for checkbox inputs.

prompt (:string) - the prompt for select inputs. Defaults to nil.

options (:list) - the options to pass to Phoenix.HTML.Form.options_for_select/2.

multiple (:boolean) - the multiple flag for select inputs. Defaults to false. Global attributes are accepted.

Slots

inner_block

label(assigns)

Renders a label.

Attributes

for (:string) - Defaults to nil.

Slots

inner_block (required)

list(assigns)

Renders a data list.

Examples

<.list>
  <:item title="Title"><%= @post.title %></:item>
  <:item title="Views"><%= @post.views %></:item>
</.list>

Slots

item (required) - Accepts attributes:

title (:string) (required)

modal(assigns)

Renders a modal.

Examples

<.modal id="confirm-modal">
  Are you sure?
  <:confirm>OK</:confirm>
  <:cancel>Cancel</:cancel>
</.modal>

JS commands may be passed to the :on_cancel and on_confirm attributes for the caller to react to each button press, for example:

<.modal id="confirm" on_confirm={JS.push("delete")} on_cancel={JS.navigate(~p"/posts")}>
  Are you sure you?
  <:confirm>OK</:confirm>
  <:cancel>Cancel</:cancel>
</.modal>

Attributes

id (:string) (required)

show (:boolean) - Defaults to false.

on_cancel (Phoenix.LiveView.JS) - Defaults to %Phoenix.LiveView.JS{ops: []}.

on_confirm (Phoenix.LiveView.JS) - Defaults to %Phoenix.LiveView.JS{ops: []}.

Slots

inner_block (required)

title

subtitle

confirm

cancel

show(js \\ %JS{}, selector)

show_modal(js \\ %JS{}, id)

simple_form(assigns)

Renders a simple form.

Examples

<.simple_form for={@form} phx-change="validate" phx-submit="save">
  <.input field={@form[:email]} label="Email"/>
  <.input field={@form[:username]} label="Username" />
  <:actions>
    <.button>Save</.button>
  </:actions>
</.simple_form>

Attributes

for (:any) (required) - the datastructure for the form.

as (:any) - the server side parameter to collect all input under. Defaults to nil. Global attributes are accepted.

Slots

inner_block (required)

actions - the slot for form actions, such as a submit button.

table(assigns)

Renders a table with generic styling.

Examples

<.table id="users" rows={@users}>
  <:col :let={user} label="id"><%= user.id %></:col>
  <:col :let={user} label="username"><%= user.username %></:col>
</.table>

Attributes

id (:string) (required)

class (:string) - Defaults to nil.

rows (:list) (required)

row_id (:any) - the function for generating the row id. Defaults to nil.

row_click (:any) - the function for handling phx-click on each row. Defaults to nil.

row_item (:any) - the function for mapping each row before calling the :col and :action slots. Defaults to &Function.identity/1.

Slots

col (required) - Accepts attributes:

label (:string)

class (:string)

action - the slot for showing user actions in the last table column.

translate_error(arg)

Translates an error message using gettext.

translate_errors(errors, field)

Translates the errors for a field from a keyword list of errors.