PlejadyWeb.UserAuth (Plejády v1.0.0)

Provides LiveView hooks and helper functions for user authentication.

Functions

fetch_current_user(conn, opts)

Fetches the user that is saved in the socket token.

get_user_by_auth(auth)

Transforms the Ueberauth.Auth struct into a User struct and saves it to the database.

This is a database action.

log_in_user(conn, user, params)

Creates a new socket and logs the user in.

log_out_user(conn)

Logs the user out.

on_mount(atom, params, session, socket)

A LiveView hook ensuring multiple levels of authentication.

The first argument is an atom and can be :ensure_authenticated, :ensure_admin or :ensure_lead.

redirect_if_user_is_authenticated(conn)

If a user is detected in the conn struct, they are automatically redirected to the app.